pyproject.toml
[tool.poetry]
name = "fastapi-fps"
version = "0.1.0"
description = ""
authors = ["Antonio Fernando <antoniofernandodj@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
dynaconf = "^3.2.4"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.23"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.24.0.post1"
mypy = "^1.7.1"
black = "^23.11.0"
pytest = "^7.4.3"
pylint = "^3.0.2"
tomli = "^2.0.1"
types-python-jose = "^3.3.4.8"
pytest-asyncio = "^0.21.1"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.0"
twisted = "^23.10.0"
aiosqlite = "^0.19.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
